APPGATE:Solving ui details..
